02_coding_test/db_backup/requirements.txt
pyodbc 
azure-storage-blob 
mysql
mysql-connector-python